[8.8] [UnifiedFieldList] Handle exceptions in FieldPopover better (#155866) (#155973)
# Backport

This will backport the following commits from `main` to `8.8`:
- [[UnifiedFieldList] Handle exceptions in FieldPopover better
(#155866)](https://github.com/elastic/kibana/pull/155866)

<!--- Backport version: 8.9.7 -->

### Questions ?
Please refer to the [Backport tool
documentation](https://github.com/sqren/backport)

<!--BACKPORT [{"author":{"name":"Julia
Rechkunova","email":"julia.rechkunova@elastic.co"},"sourceCommit":{"committedDate":"2023-04-27T08:09:56Z","message":"[UnifiedFieldList]
Handle exceptions in FieldPopover better (#155866)\n\n##
Summary\r\n\r\nThis PR wraps rendering of FieldPopover with additional
try/catch and\r\nerror boundary so we can look into exceptions if they
happen.","sha":"c571a389efaa826650172ff44d3de5f3eab3855b","branchLabelMapping":{"^v8.9.0$":"main","^v(\\d+).(\\d+).\\d+$":"$1.$2"}},"sourcePullRequest":{"labels":["release_note:skip","Team:DataDiscovery","backport:prev-minor","Feature:UnifiedFieldList","v8.9.0"],"number":155866,"url":"https://github.com/elastic/kibana/pull/155866","mergeCommit":{"message":"[UnifiedFieldList]
Handle exceptions in FieldPopover better (#155866)\n\n##
Summary\r\n\r\nThis PR wraps rendering of FieldPopover with additional
try/catch and\r\nerror boundary so we can look into exceptions if they
happen.","sha":"c571a389efaa826650172ff44d3de5f3eab3855b"}},"sourceBranch":"main","suggestedTargetBranches":[],"targetPullRequestStates":[{"branch":"main","label":"v8.9.0","labelRegex":"^v8.9.0$","isSourceBranch":true,"state":"MERGED","url":"https://github.com/elastic/kibana/pull/155866","number":155866,"mergeCommit":{"message":"[UnifiedFieldList]
Handle exceptions in FieldPopover better (#155866)\n\n##
Summary\r\n\r\nThis PR wraps rendering of FieldPopover with additional
try/catch and\r\nerror boundary so we can look into exceptions if they
happen.","sha":"c571a389efaa826650172ff44d3de5f3eab3855b"}}]}]
BACKPORT-->
