+++ src/plugins/unified_field_list/public/components/field_stats/field_stats.tsx
@@ -47,6 +47,7 @@ import {
   getDefaultColor,
 } from './field_top_values';
 import { FieldSummaryMessage } from './field_summary_message';
+import { ErrorBoundary } from '../error_boundary';
 
 export interface FieldStatsState {
   isLoading: boolean;
@@ -539,25 +540,6 @@ const FieldStatsComponent: React.FC<FieldStatsProps> = ({
 
   return null;
 };
-
-class ErrorBoundary extends React.Component<{}, { hasError: boolean }> {
-  constructor(props: FieldStatsProps) {
-    super(props);
-    this.state = { hasError: false };
-  }
-
-  static getDerivedStateFromError() {
-    return { hasError: true };
-  }
-
-  render() {
-    if (this.state.hasError) {
-      return null;
-    }
-
-    return this.props.children;
-  }
-}
 
 /**
  * Component which fetches and renders stats for a data view field
